proyecto_software/requirements.txt
PyQt6
Flask
python-dotenv
PyJWT
mysql-connector-python
pymysql
JayDeBeApi
bcrypt